+++ macaroons/index.js
@@ -1,4 +1,3 @@
-const grantAccess = require('./grant_access');
 const restrictMacaroon = require('./restrict_macaroon');
 
-module.exports = {grantAccess, restrictMacaroon};
+module.exports = {restrictMacaroon};